pyproject.toml
[project]
name = "queryRouter"
version = "0.1.0"
description = "Router di ricerca locale professionale"
dependencies = [
    "fastapi",
    "uvicorn",
    "typer",
    "pyyaml",
    "pydantic"
]

[project.scripts]
qr = "queryRouter.cli:app"